+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## WLED changelog
 
-#### Build 2403290
+#### Build 2404040
 -   WLED 0.14.3 release
 -   Fix for transition 0 (#3854, #3832, #3720)
+-   Fix for #3855 via #3873 (by @willmmiles)
 
 #### Build 2403170
 -   WLED 0.14.2 release